ml flow, transformer, and refactoring output oncverter

+++ chipiron/players/boardevaluators/neural_networks/input_converters/representation_364_bti.py
@@ -53,4 +53,4 @@ class RepresentationBTI:
         tensor: torch.Tensor = representation.get_evaluator_input(
             color_to_play=board.turn
         )
-        return tensor
+        return tensor.float()
